fixing broken tests and some refactoring
